requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djoser==2.2.3
mysqlclient==2.2.7
